+++ premium/web-frontend/modules/baserow_premium/components/PremiumFeatures.vue
@@ -20,5 +20,9 @@
       <i class="fas fa-fw fa-check premium-features__feature-icon"></i>
       {{ $t('premiumFeatures.rowColoring') }}
     </li>
+    <li class="premium-features__feature">
+      <i class="fas fa-fw fa-check premium-features__feature-icon"></i>
+      {{ $t('premiumFeatures.surveyForm') }}
+    </li>
   </ul>
 </template>