+++ web-frontend/modules/database/components/view/form/FormViewHeader.vue
@@ -2,13 +2,34 @@
   <ul v-if="!tableLoading" class="header__filter header__filter--full-width">
     <li class="header__filter-item">
       <a
+        v-if="!readOnly"
+        ref="modeContextLink"
+        class="header__filter-link"
+        @click="
+          $refs.modeContext.toggle($refs.modeContextLink, 'bottom', 'left', 4)
+        "
+      >
+        <i class="header__filter-icon fas fa-palette"></i>
+        <span class="header__filter-name">{{ $t('formViewHeader.mode') }}</span>
+      </a>
+      <FormViewModeContext
+        ref="modeContext"
+        :database="database"
+        :view="view"
+      ></FormViewModeContext>
+    </li>
+    <li class="header__filter-item">
+      <a
+        v-if="!readOnly"
         :href="formUrl"
         target="_blank"
         rel="noopener"
         class="header__filter-link"
       >
         <i class="header__filter-icon fas fa-eye"></i>
-        <span class="header__filter-name">Preview</span>
+        <span class="header__filter-name">{{
+          $t('formViewHeader.preview')
+        }}</span>
       </a>
     </li>
   </ul>
@@ -18,11 +39,17 @@
 import { mapState } from 'vuex'
 
 import formViewHelpers from '@baserow/modules/database/mixins/formViewHelpers'
+import FormViewModeContext from '@baserow/modules/database/components/view/form/FormViewModeContext'
 
 export default {
   name: 'FormViewHeader',
+  components: { FormViewModeContext },
   mixins: [formViewHelpers],
   props: {
+    database: {
+      type: Object,
+      required: true,
+    },
     view: {
       type: Object,
       required: true,